requirements.txt
base58==2.1.1
ecdsa==0.18.0
greenlet==3.0.3
pycryptodome==3.20.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
